refactor(transformer): convert transformer class to a function
