+++ src/index.ts
@@ -5,4 +5,4 @@ export * from './clause';
 export * from './clause-collection';
 export * from './clauses';
 export * from './query';
-export * from './transformer';
+export * from './transformer/transformer';
